+++ packages/web3-eth-accounts/src/common/index.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 // @ethereumjs/common version 3.1.1
-export * from './common';
-export * from './enums';
-export * from './types';
-export * from './utils';
+export * from './common.js';
+export * from './enums.js';
+export * from './types.js';
+export * from './utils.js';