+++ packages/web3-eth-accounts/src/tx/index.ts
@@ -16,8 +16,8 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 // @ethereumjs/tx version 4.1.1
-export { FeeMarketEIP1559Transaction } from './eip1559Transaction';
-export { AccessListEIP2930Transaction } from './eip2930Transaction';
-export { Transaction } from './legacyTransaction';
-export { TransactionFactory } from './transactionFactory';
-export * from './types';
+export { FeeMarketEIP1559Transaction } from './eip1559Transaction.js';
+export { AccessListEIP2930Transaction } from './eip2930Transaction.js';
+export { Transaction } from './legacyTransaction.js';
+export { TransactionFactory } from './transactionFactory.js';
+export * from './types.js';