+++ packages/web3-eth-ens/test/unit/resolver.test.ts
@@ -16,13 +16,13 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { Web3Context, Web3ContextObject } from 'web3-core';
-import { Contract, NonPayableMethodObject } from 'web3-eth-contract';
 import { ResolverMethodMissingError } from 'web3-errors';
-import { sha3Raw, sha3 } from 'web3-utils';
+import { Contract, NonPayableMethodObject } from 'web3-eth-contract';
+import { sha3 } from 'web3-utils';
+import { PublicResolverAbi } from '../../src/abi/ens/PublicResolver';
+import { interfaceIds, methodsInInterface } from '../../src/config';
 import { Registry } from '../../src/registry';
 import { Resolver } from '../../src/resolver';
-import { PublicResolverAbi } from '../../src/abi/ens/PublicResolver';
-import { methodsInInterface, interfaceIds } from '../../src/config';
 import { namehash } from '../../src/utils';
 
 describe('resolver', () => {
@@ -32,8 +32,6 @@ describe('resolver', () => {
 	let contract: Contract<typeof PublicResolverAbi>;
 	const mockAddress = '0x0000000000000000000000000000000000000000';
 	const ENS_NAME = 'web3js.eth';
-	const x = '0x1000000000000000000000000000000000000000000000000000000000000000';
-	const y = '0x2000000000000000000000000000000000000000000000000000000000000000';
 
 	beforeAll(() => {
 		const context = new Web3Context('http://test.com');
@@ -51,6 +49,13 @@ describe('resolver', () => {
 				new ResolverMethodMissingError(mockAddress, methodName),
 			);
 		});
+		it('isNullish interface with no address', async () => {
+			const methodName = 'nullish';
+			const localContract = new Contract(PublicResolverAbi);
+			await expect(resolver.checkInterfaceSupport(localContract, methodName)).rejects.toThrow(
+				new ResolverMethodMissingError('', methodName),
+			);
+		});
 
 		it('Doesn"t support interface', async () => {
 			const methodName = methodsInInterface.setAddr; // Just a method to pass first check
@@ -65,6 +70,20 @@ describe('resolver', () => {
 				new ResolverMethodMissingError(mockAddress, methodName),
 			);
 
+			expect(supportsInterfaceMock).toHaveBeenCalledWith(interfaceIds[methodName]);
+		});
+		it('Doesn"t support interface with no address', async () => {
+			const methodName = methodsInInterface.setAddr; // Just a method to pass first check
+			const localContract = new Contract(PublicResolverAbi);
+			const supportsInterfaceMock = jest
+				.spyOn(localContract.methods, 'supportsInterface')
+				.mockReturnValue({
+					call: jest.fn().mockReturnValue(false),
+				} as unknown as NonPayableMethodObject<any, any>);
+
+			await expect(resolver.checkInterfaceSupport(localContract, methodName)).rejects.toThrow(
+				new ResolverMethodMissingError('', methodName),
+			);
 			expect(supportsInterfaceMock).toHaveBeenCalledWith(interfaceIds[methodName]);
 		});
 
@@ -86,30 +105,6 @@ describe('resolver', () => {
 		);
 	});
 	describe('addr', () => {
-		it('setAddr valid', async () => {
-			const checkInteraface = jest.spyOn(resolver, 'checkInterfaceSupport');
-
-			const setAddrMock = jest.spyOn(contract.methods, 'setAddr').mockReturnValue({
-				send: jest.fn(),
-			} as unknown as NonPayableMethodObject<any, any>);
-
-			jest.spyOn(contract.methods, 'supportsInterface').mockReturnValue({
-				call: jest.fn().mockReturnValue(true),
-			} as unknown as NonPayableMethodObject<any, any>);
-
-			// todo when moving this mock in beforeAll, jest calls the actual implementation, how to fix that
-			// I use this in many places
-			jest.spyOn(registry, 'getResolver').mockImplementation(async () => {
-				return new Promise(resolve => {
-					resolve(contract);
-				});
-			});
-
-			await resolver.setAddress(ENS_NAME, mockAddress, { from: mockAddress });
-			expect(checkInteraface).toHaveBeenCalled();
-			expect(setAddrMock).toHaveBeenCalledWith(namehash(ENS_NAME), mockAddress);
-		});
-
 		it('getAddress', async () => {
 			const supportsInterfaceMock = jest
 				.spyOn(contract.methods, 'supportsInterface')
@@ -138,40 +133,6 @@ describe('resolver', () => {
 	});
 
 	describe('pubkey', () => {
-		it('setPubkey', async () => {
-			const setPubKeyMethod = methodsInInterface.setPubkey;
-
-			jest.spyOn(registry, 'getResolver').mockImplementation(async () => {
-				return new Promise(resolve => {
-					resolve(contract);
-				});
-			});
-
-			// eslint-disable-next-line @typescript-eslint/no-empty-function
-			const send = jest.spyOn({ send: () => {} }, 'send');
-
-			const setPubKeyMock = jest.spyOn(contract.methods, 'setPubkey').mockReturnValue({
-				send,
-			} as unknown as NonPayableMethodObject<any, any>);
-
-			const supportsInterfaceMock = jest
-				.spyOn(contract.methods, 'supportsInterface')
-				.mockReturnValue({
-					call: jest.fn().mockReturnValue(true),
-				} as unknown as NonPayableMethodObject<any, any>);
-
-			const sendOptions = { from: mockAddress };
-			await expect(resolver.setPubkey(ENS_NAME, x, y, sendOptions)).resolves.not.toThrow();
-
-			expect(setPubKeyMock).toHaveBeenCalledWith(
-				namehash(ENS_NAME),
-				namehash(x),
-				namehash(y),
-			);
-			expect(supportsInterfaceMock).toHaveBeenCalledWith(interfaceIds[setPubKeyMethod]);
-			expect(send).toHaveBeenCalledWith(sendOptions);
-		});
-
 		it('getPubkey', async () => {
 			const supportsInterfaceMock = jest
 				.spyOn(contract.methods, 'supportsInterface')
@@ -200,41 +161,6 @@ describe('resolver', () => {
 	});
 
 	describe('Contenthash', () => {
-		it('setContenthash', async () => {
-			const setContenthashMethod = methodsInInterface.setContenthash;
-			const hash = sha3Raw('justToHash');
-
-			jest.spyOn(registry, 'getResolver').mockImplementation(async () => {
-				return new Promise(resolve => {
-					resolve(contract);
-				});
-			});
-
-			// eslint-disable-next-line @typescript-eslint/no-empty-function
-			const send = jest.spyOn({ send: () => {} }, 'send');
-
-			const setContenthashMock = jest
-				.spyOn(contract.methods, 'setContenthash')
-				.mockReturnValue({
-					send,
-				} as unknown as NonPayableMethodObject<any, any>);
-
-			const supportsInterfaceMock = jest
-				.spyOn(contract.methods, 'supportsInterface')
-				.mockReturnValue({
-					call: jest.fn().mockReturnValue(true),
-				} as unknown as NonPayableMethodObject<any, any>);
-
-			const sendOptions = { from: mockAddress };
-			await expect(
-				resolver.setContenthash(ENS_NAME, hash, sendOptions),
-			).resolves.not.toThrow();
-
-			expect(setContenthashMock).toHaveBeenCalledWith(namehash(ENS_NAME), hash);
-			expect(supportsInterfaceMock).toHaveBeenCalledWith(interfaceIds[setContenthashMethod]);
-			expect(send).toHaveBeenCalledWith(sendOptions);
-		});
-
 		it('getContenthash', async () => {
 			const supportsInterfaceMock = jest
 				.spyOn(contract.methods, 'supportsInterface')