+++ packages/web3-eth/src/utils/try_send_transaction.ts
@@ -20,7 +20,7 @@ import { AsyncFunction, rejectIfTimeout } from 'web3-utils';
 import { TransactionSendTimeoutError } from 'web3-errors';
 
 // eslint-disable-next-line import/no-cycle
-import { rejectIfBlockTimeout } from './reject_if_block_timeout';
+import { rejectIfBlockTimeout } from './reject_if_block_timeout.js';
 
 /**
  * An internal function to send a transaction or throws if sending did not finish during the timeout during the blocks-timeout.