+++ packages/web3-providers-ipc/test/integration/reconnection.test.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import IpcProvider from '../../src';
+import { IpcProvider } from '../../src';
 
 import { startGethServer } from '../fixtures/helpers';
 import {