+++ packages/web3-eth-ens/CHANGELOG.md
@@ -83,3 +83,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Bug fix of `checkNetwork` in ENS (#5988)
 
 ## [Unreleased]
+
+### Removed
+
+-   Removed non read-only methods (#6084)